Move files to 'main' package and refactor imports and modules
