+++ src/main/__init__.py
@@ -1,0 +1,1 @@
+from .web.web import app
