+++ src/main/web/web.py
@@ -1,11 +1,11 @@
 from flask import Flask
-import submodule
+from ..submodule.hello import say_hello
 
 app = Flask(__name__)
 
 @app.route('/')
 def hello():
-    greeting = submodule.say_hello()
+    greeting = say_hello()
     return greeting
 
 def run(host='127.0.0.1', port=5000, debug=True):
