--- src/submodule/__init__.py
@@ -1,1 +1,0 @@
-from .hello import say_hello
