--- src/web/__init__.py
@@ -1,1 +1,0 @@
-from .web import run
